Trinet Layer Code/Trinet/Trinet-Layer/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "dnspython>=2.8.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
